requirements.txt
pipecat-ai
livekit
livekit-api
google-generativeai
groq
fastapi
uvicorn
redis
langchain
langchain-core
langchain-google-genai
langchain-groq
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
python-dotenv
tiktoken
pydantic
sounddevice
numpy
pypdf
python-multipart
